refactor(cli): extract config/onboarding subsystem into cli_config.py
Gets cli.py under the 10k-line-per-file budget (13,248 → 9,664). The native
subcommand extraction alone left cli.py well over budget, so move the second
large cohesive block: the interactive harness/credential configuration
subsystem behind `omnigent config` / `omnigent setup` and the first-run
`configure harnesses` picker.

- New omnigent/cli_config.py (~3,650 lines) holds the 63 config helpers:
  _configure_harness_add, every _manage_*_harness / _prompt_install_* / _set_*,
  the ambient-credential adoption path, node-dependency preflight, and
  _run_configure_harnesses_interactive. _CLI_LOGIN_BRAND moves with them (it had
  no other user). The config/setup/integration Click commands stay in cli.py.
- The 3 config-load helpers the block needs (_load_global_config /
  _save_global_config / _load_effective_config) stay in cli.py (used ~20x each
  there); cli_config reaches them through call-time proxies, so importing
  cli_config never imports omnigent.cli (no cycle) and monkeypatching
  omnigent.cli.<helper> is still honoured.
- cli.py re-imports the 7 config entry points its commands call, so they remain
  omnigent.cli attributes (patchable, importable) for callers and tests.
- Tests: repoint references for helpers that are called *intra*-cli_config to
  omnigent.cli_config (where patching now takes effect) — the _manage_* dispatch
  test, _adopt_detected_providers / _promote_global_auth_to_provider /
  _launch_*_configure / _qwen_auth_configured patches, and the opencode / promote
  imports. Helpers cli.py itself calls stay patched on omnigent.cli.

Behavior-preserving; no command, flag, or prompt changed.

Test plan: tests/cli/{test_cli,test_configure_models,test_opencode_setup,
test_chat,test_import,test_backend,test_runner_startup}.py all pass; ruff
format+check and pre-commit file hooks clean; cli.py is 9,664 lines.
