+++ tests/cli/test_opencode_setup.py
@@ -9,7 +9,9 @@ from types import SimpleNamespace
 import pytest
 
 import omnigent.cli as cli
-from omnigent.cli import _list_opencode_models, _load_global_config, _set_opencode_default_model
+import omnigent.cli_config as cli_config
+from omnigent.cli import _load_global_config
+from omnigent.cli_config import _list_opencode_models, _set_opencode_default_model
 
 
 @pytest.fixture
@@ -32,7 +34,7 @@ def test_list_models_parses_nonblank_lines(monkeypatch: pytest.MonkeyPatch) -> N
         "omnigent.onboarding.harness_install.harness_install_spec", lambda _key: _fake_spec()
     )
     monkeypatch.setattr(
-        cli.subprocess,
+        cli_config.subprocess,
         "run",
         lambda *a, **k: subprocess.CompletedProcess(
             a, 0, stdout="anthropic/claude-sonnet-4-5\nopenai/gpt-5.5\n\n  \n", stderr=""
@@ -56,7 +58,7 @@ def test_list_models_empty_on_subprocess_error(monkeypatch: pytest.MonkeyPatch) 
     def _boom(*_a: object, **_k: object) -> object:
         raise OSError("no binary")
 
-    monkeypatch.setattr(cli.subprocess, "run", _boom)
+    monkeypatch.setattr(cli_config.subprocess, "run", _boom)
     assert _list_opencode_models() == []
 
 
@@ -67,7 +69,7 @@ def test_set_default_model_persists_choice(
     monkeypatch: pytest.MonkeyPatch, _isolated_config: Path
 ) -> None:
     monkeypatch.setattr(
-        cli, "_list_opencode_models", lambda: ["anthropic/claude-sonnet-4-5", "x/y"]
+        cli_config, "_list_opencode_models", lambda: ["anthropic/claude-sonnet-4-5", "x/y"]
     )
     monkeypatch.setattr("omnigent.onboarding.interactive.select", lambda *a, **k: 0)
     status = _set_opencode_default_model(current=None)
@@ -79,7 +81,7 @@ def test_set_default_model_clear_unsets(
     monkeypatch: pytest.MonkeyPatch, _isolated_config: Path
 ) -> None:
     cli._save_global_config({"opencode_model": "x/y"})
-    monkeypatch.setattr(cli, "_list_opencode_models", lambda: ["a/b"])
+    monkeypatch.setattr(cli_config, "_list_opencode_models", lambda: ["a/b"])
     # options == ["a/b", "Clear default ..."]; index 1 is the clear row.
     monkeypatch.setattr("omnigent.onboarding.interactive.select", lambda *a, **k: 1)
     status = _set_opencode_default_model(current="x/y")
@@ -90,14 +92,14 @@ def test_set_default_model_clear_unsets(
 def test_set_default_model_cancel_is_noop(
     monkeypatch: pytest.MonkeyPatch, _isolated_config: Path
 ) -> None:
-    monkeypatch.setattr(cli, "_list_opencode_models", lambda: ["a/b"])
+    monkeypatch.setattr(cli_config, "_list_opencode_models", lambda: ["a/b"])
     monkeypatch.setattr("omnigent.onboarding.interactive.select", lambda *a, **k: -1)
     assert _set_opencode_default_model(current=None) is None
     assert _load_global_config() == {}
 
 
 def test_set_default_model_no_models_short_circuits(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(cli, "_list_opencode_models", list)
+    monkeypatch.setattr(cli_config, "_list_opencode_models", list)
     called = False
 
     def _select(*_a: object, **_k: object) -> int:
